move the api view from view.py to api_view.py
