+++ hackathon_site/event/api_views.py
@@ -1,15 +1,21 @@
 from django.db import transaction
 from django.db.models import Q
+from django.shortcuts import get_object_or_404
+from django_filters import rest_framework as filters
 
 from rest_framework import generics, mixins, status
 from rest_framework.exceptions import ValidationError
 from rest_framework.response import Response
+from rest_framework.filters import SearchFilter
+
 
 from event.models import User, Team as EventTeam
 from event.serializers import UserSerializer, TeamSerializer
-from event.permissions import UserHasProfile
+from event.permissions import UserHasProfile, FullDjangoModelPermissions
+from event.api_filters import TeamFilter
 
 from hardware.models import OrderItem
+
 
 
 class CurrentUserAPIView(generics.GenericAPIView, mixins.RetrieveModelMixin):
@@ -101,3 +107,43 @@ class JoinTeamView(generics.GenericAPIView, mixins.RetrieveModelMixin):
         response_serializer = TeamSerializer(profile.team)
         response_data = response_serializer.data
         return Response(data=response_data, status=status.HTTP_200_OK,)
+
+
+class MultipleFieldLookupMixin:
+    """
+    Apply this mixin to any view or viewset to get multiple field filtering
+    based on a `lookup_fields` attribute, instead of the default single field filtering.
+    """
+
+    def get_object(self, *args, **kwargs):
+        queryset = self.get_queryset()  # Get the base queryset
+        queryset = self.filter_queryset(queryset)  # Apply any filter backends
+        filter = {}
+        for field in self.lookup_fields:
+            if self.kwargs[field]:  # Ignore empty fields.
+                filter[field] = self.kwargs[field]
+        obj = get_object_or_404(queryset, **filter)  # Lookup the object
+        self.check_object_permissions(self.request, obj)
+        return obj
+
+
+class TeamCodeView(MultipleFieldLookupMixin, generics.GenericAPIView):
+    queryset = EventTeam.objects.all()
+    serializer_class = TeamSerializer
+    permission_classes = [FullDjangoModelPermissions]
+    lookup_fields = ["team_code"]
+
+    filter_backends = (filters.DjangoFilterBackend, SearchFilter)
+    filterset_class = TeamFilter
+    search_fields = (
+        "team_code",
+        "id",
+        "profiles__user__first_name",
+        "profiles__user__last_name",
+    )
+
+    def get(self, request, *args, **kwargs):
+        return Response(
+            TeamSerializer(self.get_object(request, *args, **kwargs)).data,
+            status=status.HTTP_200_OK,
+        )
